filemap.js: No longer bail when file/journal IDs mismatch - always use journal.
During merge, copy the journalmap so we don't modify the original data structure.

journal-sync.js: Test returns to also showing the file and journal maps before the merge happens.
Computed actions now includes extra data to aid in stacked folder creation.
Added a mkfolder action.
Added depth into all acitons.
Added util to find a journal folder by name.
Added doActionMkFolder to create journal folders.
When importing journals, if no journal in the action, assume it was created, and search for it.
Convert forEach loops for directories and folder actions and repalce with for loops so
we can await for them to finish import/exports depend on them.